requirements.txt
absl-py==0.12.0
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dataclasses==0.8
google-auth==1.28.0
google-auth-oauthlib==0.4.3
grpcio==1.36.1
h5py==3.1.0
idna==2.10
importlib-metadata==3.7.3
kiwisolver==1.3.1
llvmlite==0.36.0
Markdown==3.3.4
matplotlib==3.3.4
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.0
pandas==1.1.5
Pillow==8.1.2
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.1
torch==1.8.0
torchaudio==0.8.0
torchvision==0.9.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
fair-esm